pyproject.toml
[project]
name = "project4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.22",
    "numpy>=1.26.0",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.2.0",
    "streamlit>=1.32.0",
]
